// src/app/type.d.ts
export type Action<Type, Payload> = { type: Type; payload: Payload };

export type MaterialConfigId = string;
export type FinishGroupId = string;
export type MaterialGroupId = string;
export type MaterialId = string;
export type QuoteId = string;
export type VendorId = string;
export type ConfigId = string;
export type ShippingId = string;
export type ModelId = string;
export type ModelSceneId = string;
export type FileId = string;
export type PriceId = string;
export type CartId = string;
export type ConfigurationId = string;
// ...
export type Location = {
  city: string;
  zipCode: string;
  stateCode: string;
  countryCode: string;
};
// ...
export enum ModalContentType {
  PICK_LOCATION= 'PICK_LOCATION',
  PICK_UNIT= 'PICK_UNIT',
  MODEL_VIEWER= 'MODEL_VIEWER',
  MATERIAL= 'MATERIAL',
  FINISH_GROUP= 'FINISH_GROUP',
  CONFIRM_LOCATION_CHANGE= 'CONFIRM_LOCATION_CHANGE',
  CONFIRM_CURRENCY_CHANGE= 'CONFIRM_CURRENCY_CHANGE',
  SHARE_CONFIGURATION= 'SHARE_CONFIGURATION',
  FATAL_ERROR= 'FATAL_ERROR',
  ERROR= 'ERROR'
}

export type ModalConfigClosed = null;
export type ModalConfigOpened = {
  isCloseable: boolean;
  contentType: ModalContentType;
  contentProps: any;
};